Changing the the approach to the problem
Initially I was going to use a node along with generics for multiple types, but began to struggle on what methods I needed in which. While writing the Node class, I found it easier to implement a one class to suit the request. It will only allow for integers, but gave me a better understanding at solving the problem rather than struggling on coding efficiency/capability.