I saw Larry Page as an inspirational software Engineer
As secretive as Larry Page is to finding details of his personal life, one of the huge difficulties in this was trying to condense all of the findings about his studies and career life. Larry Page did not seem like your typical business man, but his drive and visions were exceptional. He became more comfortable in the later lift with the managing of the companies. He remains to be imaginative, creative and inventive even to this day. It was through these attributes that made him inspirational. It was the Google search engine that led me to writing about him.